prima: Get timestamp in gscan_start and add it to cached results
Firmware gives the time from gscan_start to the time at which a
scan result is found as timestamp of that result. Get the system
timestamp in gscan_start in driver and add it to the timstamp
provided by firmware as framework expects the timestamp of the
cached result from system boot.

Change-Id: I7b1f7f671558daedb7513eb797b7bb918309fe8a
CRs-Fixed: 996615